refactoring: moved product, image and category serializers to different files
