+++ apps/products/serializers/__init__.py
@@ -1,0 +1,3 @@
+from .categories import CategorySerializer  # noqa
+from .images import ImageSerializer  # noqa
+from .products import ProductReadSerializer, ProductWriteSerializer  # noqa
