+++ apps/products/serializers/images.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from apps.products.models import Image
+
+
+class ImageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Image
+        fields = ("image",)
